requirements.txt
nltk
spacy
pdfminer.six
streamlit
